#131: Rename gps-related modules to reduce nomenclature confusion
