+++ track/__init__.py
@@ -6,4 +6,4 @@ from .errorsources import *
 from .telem import *
 from .laser import *
 from .plate_solve import *
-from .gpsd import *
+from .gps_client import *
